--- backend/test/BUILD.bazel
@@ -1,54 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["test_utils.go"],
-    importpath = "github.com/kubeflow/pipelines/backend/test",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//backend/api/go_http_client/experiment_client/experiment_service:go_default_library",
-        "//backend/api/go_http_client/job_client/job_service:go_default_library",
-        "//backend/api/go_http_client/pipeline_client/pipeline_service:go_default_library",
-        "//backend/api/go_http_client/run_client/run_service:go_default_library",
-        "//backend/src/common/client/api_server:go_default_library",
-        "@com_github_cenkalti_backoff//:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_stretchr_testify//assert:go_default_library",
-        "@io_k8s_client_go//tools/clientcmd:go_default_library",
-        "@io_k8s_client_go//tools/clientcmd/api:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = [
-        "cli_test.go",
-        "experiment_api_test.go",
-        "job_api_test.go",
-        "pipeline_api_test.go",
-        "run_api_test.go",
-    ],
-    embed = [":go_default_library"],
-    deps = [
-        "//backend/api:go_default_library",
-        "//backend/api/go_http_client/experiment_client/experiment_service:go_default_library",
-        "//backend/api/go_http_client/experiment_model:go_default_library",
-        "//backend/api/go_http_client/job_client/job_service:go_default_library",
-        "//backend/api/go_http_client/job_model:go_default_library",
-        "//backend/api/go_http_client/pipeline_client/pipeline_service:go_default_library",
-        "//backend/api/go_http_client/pipeline_model:go_default_library",
-        "//backend/api/go_http_client/pipeline_upload_client/pipeline_upload_service:go_default_library",
-        "//backend/api/go_http_client/run_client/run_service:go_default_library",
-        "//backend/api/go_http_client/run_model:go_default_library",
-        "//backend/src/cmd/ml/cmd:go_default_library",
-        "//backend/src/common/client/api_server:go_default_library",
-        "//backend/src/common/util:go_default_library",
-        "@com_github_argoproj_argo//pkg/apis/workflow/v1alpha1:go_default_library",
-        "@com_github_ghodss_yaml//:go_default_library",
-        "@com_github_golang_glog//:go_default_library",
-        "@com_github_stretchr_testify//assert:go_default_library",
-        "@com_github_stretchr_testify//suite:go_default_library",
-        "@io_k8s_apimachinery//pkg/util/yaml:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-    ],
-)
