+++ backend/test/integration_test/BUILD.bazel
@@ -1,0 +1,20 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["test_utils.go"],
+    importpath = "github.com/kubeflow/pipelines/backend/test/integration_test",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//backend/api/go_http_client/experiment_client/experiment_service:go_default_library",
+        "//backend/api/go_http_client/job_client/job_service:go_default_library",
+        "//backend/api/go_http_client/pipeline_client/pipeline_service:go_default_library",
+        "//backend/api/go_http_client/run_client/run_service:go_default_library",
+        "//backend/src/common/client/api_server:go_default_library",
+        "@com_github_cenkalti_backoff//:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
+        "@io_k8s_client_go//tools/clientcmd:go_default_library",
+        "@io_k8s_client_go//tools/clientcmd/api:go_default_library",
+    ],
+)
