+++ backend/test/integration_test/cli_test.go
@@ -1,4 +1,4 @@
-package test
+package integration_test
 
 import (
 	"testing"
