+++ backend/test/integration_test/pipeline_api_test.go
@@ -1,4 +1,4 @@
-package test
+package integration_test
 
 import (
 	"io/ioutil"
