+++ backend/test/integration_test/test_utils.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package test
+package integration_test
 
 import (
 	"fmt"
